data_scraping_pipeline/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
loguru==0.7.3
pydantic-settings==2.7.1
gunicorn==23.0.0
python-jose[cryptography]==3.3.0
langchain==0.3.23
langchain_google_genai==2.1.2
langchain-openai==0.3.12
langchain-voyageai==0.1.4
langchain-community>=0.3.21
langchain-mongodb>=0.2.0
pymongo>=4.6.0
ultimate-sitemap-parser>=1.3.1
apscheduler>=3.11.0